+++ src/main/java/com/koitourdemo/demo/model/response/KoiResponse.java
@@ -4,7 +4,6 @@ import lombok.AccessLevel;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -15,7 +14,7 @@ public class KoiResponse {
     String type;
     String size;
     String origin;
-    BigDecimal price;
+    float price;
     String image;
     Date createAt;
 }
